Parrot

A neural network framework for emulating stellar population synthesis. Written in Julia.

Installation

First, please install Julia on your system by following the steps listed on the Julia website. This package is currently being developed and tested on Julia v1.5, but it will likely work with earlier versions as well.

To install Parrot, open a Julia REPL and type ] (to enter the Pkg REPL) followed by:

(@v1.5) pkg> add https://github.com/elijahmathews/Parrot.jl.git

Or alternatively, run the following code directly at the Julia REPL:

julia> import Pkg; Pkg.add(url="https://github.com/elijahmathews/Parrot.jl.git")

Overview

At the moment, Parrot is in development, but the intent is to create a package that will aid in emulating stellar population synthesis (SPS) codes using neural networks. The package is currently being designed under the Flux machine learning ecosystem.

Parrot.Alsing — Type

Alsing(in::Integer, out::Integer)

Create a non-linear Alsing layer with trainable parameters W, b, α, and β.

y = (β .+ σ.(α .* (W*x .+ b)) .* (1 .- β)) .* (W*x .+ b)

The input x must be a vector of length in, or a batch of vectors represented as an in × N matrix. The out y will be a vector or batch of length out.

See arXiv:1911.11778 for more information about this activation function. Intended to be used for emulating stellar population synthesis codes.

Example

julia> a = Alsing(8,3)
Alsing(8, 3)

julia> a(rand(8))
3-element Array{Float64,1}:
  0.9555069887698046
  0.14443222298921962
 -0.06765412728860333

Parrot.ReconstructPCA — Type

ReconstructPCA(P::MultivariateStats.PCA{AbstractFloat})

Create a ReconstructPCA layer that uses Principal Component Analysis (PCA) to reconstruct its input using a given MultivariateStats.PCA object P.

y = MultivariateStats.reconstruct(P, x)

The input x must have dimensions compatible with P and the layer's output will be of the dimensions given by the PCA reconstruction. No parameters are trainable.

See also: MultivariateStats.PCA, MultivariateStats.reconstruct

Parrot.Normalization — Type

Normalization(μ::AbstractArray, σ::AbstractArray)

Create a simple Normalization layer with parameters consisting of some previously known mean μ and standard deviation σ that are used to normalize its input.

y = (x .- μ) ./ σ

The input x must be a vector of equal length to both μ and σ or an array with dimensions such that the broadcasted functions can be used with the given μ and σ. No parameters are trainable.

Parrot.Denormalization — Type

Denormalization(μ::AbstractArray, σ::AbstractArray)

Create a simple Denormalization layer with parameters consisting of some previously known mean μ and standard deviation σ that are used to restore its input from a normalized state.

y = (σ .* x) .+ μ

The input x must be a vector of equal length to both μ and σ or an array with dimensions such that the broadcasted functions can be used with the given μ and σ. No parameters are trainable.